+++ rtl/sys_gen/w11a/basys3/sys_w11a_b3.vhd
@@ -1,4 +1,4 @@
--- $Id: sys_w11a_b3.vhd 1055 2018-10-12 17:53:52Z mueller $
+-- $Id: sys_w11a_b3.vhd 1056 2018-10-13 16:01:17Z mueller $
 --
 -- Copyright 2015-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
@@ -36,6 +36,7 @@
 --
 -- Synthesized:
 -- Date         Rev  viv    Target       flop  lutl  lutm  bram  slic
+-- 2018-10-13  1055 2017.2  xc7a35t-1    2698  5636   170  47.5  1723 +dmpcnt
 -- 2018-09-15  1045 2017.2  xc7a35t-1    2475  5282   138  47.5  1643 +KW11P
 -- 2017-04-16   881 2016.4  xc7a35t-1    2412  5228   138  47.5  1608 +DEUNA
 -- 2017-01-29   846 2016.4  xc7a35t-1    2362  5239   138  47.5  1619 +int24
@@ -50,7 +51,7 @@
 --
 -- Revision History: 
 -- Date         Rev Version  Comment
--- 2018-10-07  1054   2.4    use DM_STAT_EXP
+-- 2018-10-13  1055   2.4    use DM_STAT_EXP; IDEC to maxisys; setup PERFEXT
 -- 2016-04-02   758   2.3.1  add rbd_usracc (bitfile+jtag timestamp access)
 -- 2016-03-28   755   2.3    use serport_2clock2
 -- 2016-03-19   748   2.2.2  define rlink SYSID
@@ -312,14 +313,14 @@ begin
       SER_MONI => SER_MONI
     );
 
-  PERFEXT(0) <= '0';
-  PERFEXT(1) <= '0';
-  PERFEXT(2) <= '0';
-  PERFEXT(3) <= '0';
-  PERFEXT(4) <= '0';
-  PERFEXT(5) <= '0';
-  PERFEXT(6) <= '0';
-  PERFEXT(7) <= CE_USEC;
+  PERFEXT(0) <= '0';                    -- unused (ext_rdrhit)
+  PERFEXT(1) <= '0';                    -- unused (ext_wrrhit)
+  PERFEXT(2) <= '0';                    -- unused (ext_wrflush)
+  PERFEXT(3) <= SER_MONI.rxact;         -- ext_rlrxact
+  PERFEXT(4) <= not SER_MONI.rxok;      -- ext_rlrxback
+  PERFEXT(5) <= SER_MONI.txact;         -- ext_rltxact
+  PERFEXT(6) <= not SER_MONI.txok;      -- ext_rltxback
+  PERFEXT(7) <= CE_USEC;                -- ext_usec
 
   SYS70 : pdp11_sys70                   -- 1 cpu system ----------------------
     port map (
@@ -359,6 +360,7 @@ begin
       RESET    => GRESET,
       BRESET   => BRESET,
       ITIMER   => DM_STAT_EXP.se_itimer,
+      IDEC     => DM_STAT_EXP.se_idec,
       CPUSUSP  => CP_STAT.cpususp,
       RB_LAM   => RB_LAM(15 downto 1),
       IB_MREQ  => IB_MREQ,