+++ rtl/sys_gen/w11a/cmoda7/sys_w11a_c7.vhd
@@ -1,4 +1,4 @@
--- $Id: sys_w11a_c7.vhd 1055 2018-10-12 17:53:52Z mueller $
+-- $Id: sys_w11a_c7.vhd 1056 2018-10-13 16:01:17Z mueller $
 --
 -- Copyright 2017-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
@@ -38,6 +38,7 @@
 --
 -- Synthesized:
 -- Date         Rev  viv    Target       flop  lutl  lutm  bram  slic
+-- 2018-10-13  1055 2017.2  xc7a35t-1    3107  6215   182  50.0  1889 +dmpcnt
 -- 2018-09-15  1045 2017.2  xc7a35t-1    2883  5891   150  50.0  1826 +KW11P
 -- 2017-06-27   918 2017.1  xc7a35t-1    2823  5827   150  50.0  1814 16kB cache
 -- 2017-06-25   916 2017.1  xc7a35t-1    2823  5796   150  47.5  1744 +BRAM
@@ -45,7 +46,7 @@
 --
 -- Revision History: 
 -- Date         Rev Version  Comment
--- 2018-10-07  1054   1.2    use DM_STAT_EXP
+-- 2018-10-13  1055   1.2    use DM_STAT_EXP; IDEC to maxisys; setup PERFEXT
 -- 2017-06-27   918   1.1.1  use 16 kB cache (all BRAM's used up)
 -- 2017-06-25   916   1.1    add bram_memctl for 672 kB total memory
 -- 2017-06-24   914   1.0    Initial version (derived from sys_w11a_n4)
@@ -284,14 +285,14 @@ begin
       SER_MONI => SER_MONI
     );
 
-  PERFEXT(0) <= '0';
-  PERFEXT(1) <= '0';
-  PERFEXT(2) <= '0';
-  PERFEXT(3) <= '0';
-  PERFEXT(4) <= '0';
-  PERFEXT(5) <= '0';
-  PERFEXT(6) <= '0';
-  PERFEXT(7) <= CE_USEC;
+  PERFEXT(0) <= '0';                    -- unused (ext_rdrhit)
+  PERFEXT(1) <= '0';                    -- unused (ext_wrrhit)
+  PERFEXT(2) <= '0';                    -- unused (ext_wrflush)
+  PERFEXT(3) <= SER_MONI.rxact;         -- ext_rlrxact
+  PERFEXT(4) <= not SER_MONI.rxok;      -- ext_rlrxback
+  PERFEXT(5) <= SER_MONI.txact;         -- ext_rltxact
+  PERFEXT(6) <= not SER_MONI.txok;      -- ext_rltxback
+  PERFEXT(7) <= CE_USEC;                -- ext_usec
   
   SYS70 : pdp11_sys70                   -- 1 cpu system ----------------------
     port map (
@@ -331,6 +332,7 @@ begin
       RESET    => GRESET,
       BRESET   => BRESET,
       ITIMER   => DM_STAT_EXP.se_itimer,
+      IDEC     => DM_STAT_EXP.se_idec,
       CPUSUSP  => CP_STAT.cpususp,
       RB_LAM   => RB_LAM(15 downto 1),
       IB_MREQ  => IB_MREQ,